Initial cleanup of the debug statements (qDebug, dbk, dbe)
Summary:
There was a combination of qDebug, kDebug and custom macros
which redirected to qDebug, standard error or standard output.

This is the first attempt to unify all of the above with the usage
of Qt categorized logging. Fixed in the patch:
- direct usage of qDebug;
- kDebug though the dbk macro;
- standard error through the dbe macro.
The levels for dbk/dbe messages have been aligned with the level
of the respective macros.
Few standard output/dbo macros have been cleaned as well, but many
are still used. At least their usage is confined to one file only
(src/generator/sudokuboard.cpp). Some of those messages produces
formatted output, so they may require overloads of the << operator,
if they need to be kept.

Test Plan: Compiles.

Reviewers: ouwerkerk, stikonas, #kde_games

Reviewed By: stikonas

Subscribers: #kde_games

Differential Revision: https://phabricator.kde.org/D5626