Porting fix: no values can be added to a null QTime
The time shown at the end of the game is always 0 seconds otherwise.
Properly initialize the QTime instance so that it is not a null value.
The C++ API changes for Qt5 states that "Adding days to a null QDate
or seconds to a null QTime will no longer return a valid QDate/QTime."