Fix 25x25 letter markers not showing up for ksudoku_scrible.svg
KSudoku allows to mark cells with candidate numbers/letters before
entering the correct answer, however for the 25x25 game variant in the
Scribble theme nothing would show up. In addition, for any game variant
using the Scribble theme lots of warnings like the following would be
printed:

> qt.svg: [...]/ksudoku_scrible.svg:7900:20: Could not add child element
> to parent element because the types are incorrect.

This is due to a whole `<g>` group being illegally inserted into the
`<path>` with `id="path3500"`. This can be confirmed with
https://validator.w3.org/check:

> SVG element g not allowed as child of SVG element path in this
> context.

By properly closing the `<path />` and moving `<g>` one level up, there
are no more warnings and the cell markers start to show up again.

Note that the SVG did not change after 25x25 cell markers where
introduced in the Qt4-based 5a65ba1459, where they worked properly (as
confirmed in a VM) despite the non-conformant SVG. It is likely that
only after a change to Qt5 itself this stopped working.

Test Plan:
  - Switch to the Scribble theme. No warnings should be printed.
  - Start a 25x25 Sudoku game, still using the Scribble theme. Add cell
  markers (e.g. via Shift + letter key or right click) and confirm they
  show up.