Correct icon and text positions in game chooser for HiDPI mode
KSudoku's welcome screen features a game chooser with a list including
icons and text for each game variant. In HiDPI mode, icons would be
positioned too far to the top, and text would get shifted to the right.

This is caused by not taking into account the increased device pixel
dimensions of the HiDPI icon when calculating icon position and text
offset and can be fixed correspondingly.

Test Plan:
Run `QT_SCALE_FACTOR=1.9 ksudoku`. In the list of game variants, check
that icons are centered vertically and that there is no excessive gap
left to the text.